Action structs get their own file, diff updates to behave like our other diffs
